% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.R
\docType{data}
\name{img}
\alias{img}
\title{default image}
\format{This (and the next) is a new tag that applies for data documentation.}
\source{
Note that you shouldn't use the `@export` tag when documenting data!
}
\usage{
img
}
\description{
Self-explanatory!
}
\keyword{datasets}
