% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2luma.R
\name{rgb2luma}
\alias{rgb2luma}
\title{Convert rgb image in png format into luma image
Needs library(png)}
\usage{
rgb2luma(img_png = TRUE)
}
\arguments{
\item{img_png}{The name of the png impage file in the directory to be converted}
}
\value{
A matrix that is the sum of three channeles of the png file (R, G, and B) and normalization (for \code{rgb2luma}).
}
\description{
Convert rgb image in png format into luma image
Needs library(png)
}
\details{
This function isn't complicated.

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item These functions are quite simple.
     \item There's nothing else to say about these functions.
}
function that these functions depend on.
}
\examples{
rgb2luma()
}
